refactor: 컨트롤러 메소드 구조 수정
