+++ src/main/java/blackjack/dto/DealResultDto.java
@@ -1,4 +1,4 @@
-package blackjack.dto.response;
+package blackjack.dto;
 
 import blackjack.domain.Dealer;
 import blackjack.domain.Players;
